+++ php/core/definitions.php
@@ -17,8 +17,8 @@ function postworld_definitions(){
 	 * Postworld Info
 	 */
 	$GLOBALS['pw']['info'] = array(
-		'version'		=>	1.605,
-		'db_version'	=>	1.30,
+		'version'		=>	'1.7.1',
+		'db_version'	=>	'1.32',
 		'mode'	=>	pw_mode(),
 		'slug'	=>	'postworld',
 		);
@@ -76,12 +76,15 @@ function postworld_definitions(){
 	 * Define the META keys.
 	 * Used in 'wp_postmeta' and 'pw_termmeta' tables
 	 */
-	define( 'PW_POSTMETA_KEY',	$theme_slug.'_meta', 		true ); // Case in-sensitive
-	define( 'PW_USERMETA_KEY',	$theme_slug.'_meta', 		true ); // Case in-sensitive
-	define( 'PW_TAXMETA_KEY',	$theme_slug.'_meta', 		true ); // Case in-sensitive
-	define( 'PW_AVATAR_KEY',	$theme_slug.'_avatar', 		true ); // Case in-sensitive
-	define( 'PW_COLORS_KEY',	$theme_slug.'_colors', 		true ); // Case in-sensitive
-	define( 'PW_EVENT_KEY',		$theme_slug.'_event', 		true ); // Case in-sensitive
+	define( 'PW_POSTMETA_KEY',		$theme_slug.'_meta', 		true ); // Case in-sensitive
+	define( 'PW_USERMETA_KEY',		$theme_slug.'_meta', 		true ); // Case in-sensitive
+	define( 'PW_TAXMETA_KEY',		$theme_slug.'_meta', 		true ); // Case in-sensitive
+	define( 'PW_AVATAR_KEY',		$theme_slug.'_avatar', 		true ); // Case in-sensitive
+	define( 'PW_COLORS_KEY',		$theme_slug.'_colors', 		true ); // Case in-sensitive
+	define( 'PW_EVENT_KEY',			$theme_slug.'_event', 		true ); // Case in-sensitive
+	define( 'PW_LINK_URL_KEY',		$theme_slug.'_link_url', 	true ); // Case in-sensitive
+	define( 'PW_LINK_FORMAT_KEY',	$theme_slug.'_link_format', true ); // Case in-sensitive
+
 
 	/**
 	 * Define the OPTIONS keys.