+++ php/core/install.php
@@ -25,8 +25,6 @@ function postworld_install(){
 				post_id BIGINT(20) unsigned NOT NULL,
 				author_id BIGINT(20) UNSIGNED NOT NULL,
 				post_class char(16) NOT NULL,
-				link_format char(16) NOT NULL,
-				link_url varchar(512) DEFAULT '' NOT NULL,
 				post_points mediumint(10) DEFAULT '0' NOT NULL,
 				rank_score mediumint(10) DEFAULT '0' NOT NULL,
 				post_shares mediumint(10) DEFAULT '0' NOT NULL,